added option to share files if type is public

+++ database/migrations/2023_12_03_084152_create_files_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignUuid('user_id')->constrained()->cascadeOnDelete();
             $table->string('name');
             $table->string('storedName');
-            $table->float('size');
+            $table->double('size');
             $table->enum('type', ['public', 'private'])->default('private');
             $table->timestamps();
         });
